Increase the minimum delay between 2 drill-down
During tests, it happend very often that a mouse scroll launched 2 drill down. This is due to the fact that the browser is busy and delay part of the multiples mouse scrolls (because we never scroll by one unit), probably because of the render of the SVG, and finally a scroll manage to pass the 700ms previous limit. With 3s, it's much better and none should drill down twice in 3s.